Implemented code for #642
Combined duplicate Base Classes into one class/file:
1.  TabBaseViewModel and ProTabBaseViewMode
2.  OutputCoordinateViewModel and ProOutputCoordinateViewModel
