+++ source/CoordinateConversion/ProAppCoordConversionModule/Config.daml
@@ -6,7 +6,7 @@
     <Image>Images\AddinDesktop32.png</Image>
     <Author>Esri</Author>
     <Company>Esri</Company>
-    <Date>3/7/2020</Date>
+    <Date>3/11/2020</Date>
     <Subject>Framework</Subject>
     <!-- Note subject can be one or more of these topics:
                     Content, Framework, Editing, Geodatabase, Geometry, Geoprocessing, Layouts, Map Authoring, Map Exploration -->
